+++ src/commonTest/kotlin/operations/logic/EqualsTest.kt
@@ -1,4 +1,4 @@
-package operations.math
+package operations.logic
 
 import TestInput
 import io.kotest.core.spec.style.FunSpec
@@ -6,14 +6,14 @@ import io.kotest.datatest.withData
 import io.kotest.matchers.shouldBe
 
 // TODO extend unit tests cases
-class NotEqualsTest : FunSpec({
-    context("JsonLogic evaluation with NotEquals operation") {
+class EqualsTest : FunSpec({
+    context("JsonLogic evaluation with Equals operation") {
         withData(
             // given
             listOf(
-                TestInput(mapOf("!=" to listOf(1, 1)), null, false),
-                TestInput(mapOf("!=" to listOf(1, "1")), null, false),
-                TestInput(mapOf("!=" to listOf(1, 2)), null, true)
+                TestInput(mapOf("==" to listOf(1, 1)), null, true),
+                TestInput(mapOf("==" to listOf(1, "1")), null, true),
+                TestInput(mapOf("==" to listOf(1, 2)), null, false)
             )
         ) { (expression, data, result) ->
             // when
